fix more tests, need to also manage cases with hive partitioning
