+++ src/storage/serialization/serialize_table_filter.cpp
@@ -10,6 +10,7 @@
 #include "duckdb/planner/filter/constant_filter.hpp"
 #include "duckdb/planner/filter/conjunction_filter.hpp"
 #include "duckdb/planner/filter/struct_filter.hpp"
+#include "duckdb/planner/filter/zone_map_filter.hpp"
 
 namespace duckdb {
