+++ src/storage/table/row_group.cpp
@@ -21,6 +21,7 @@
 #include "duckdb/common/serializer/binary_serializer.hpp"
 #include "duckdb/planner/filter/conjunction_filter.hpp"
 #include "duckdb/planner/filter/struct_filter.hpp"
+#include "duckdb/planner/filter/zone_map_filter.hpp"
 #include "duckdb/execution/adaptive_filter.hpp"
 
 namespace duckdb {
